+++ blueprints/managers.py
@@ -4,6 +4,7 @@ from typing import Tuple
 from bravado.exception import HTTPForbidden, HTTPUnauthorized
 from django.contrib.auth.models import User
 from django.db import models
+from django.db.models import Q
 from django.utils.timezone import now
 
 from allianceauth.services.hooks import get_extension_logger
@@ -228,30 +229,53 @@ class LocationManager(models.Manager):
         )
 
 
+class RequestQuerySet(models.QuerySet):
+    def requests_fulfillable_by_user(self, user: User) -> models.QuerySet:
+        ownerships = user.character_ownerships.all()
+        corporation_ids = {
+            character.character.corporation_id for character in ownerships
+        }
+        character_ownership_ids = {character.pk for character in ownerships}
+        request_query = self.select_related(
+            "blueprint__owner", "blueprint__owner__corporation"
+        ).filter(
+            (
+                Q(blueprint__owner__corporation__corporation_id__in=corporation_ids)
+                | Q(blueprint__owner__pk__in=character_ownership_ids)
+            )
+            & Q(closed_at=None)
+            & Q(status="OP")
+        )
+        return request_query
+
+    def requests_being_fulfilled_by_user(self, user: User) -> models.QuerySet:
+        ownerships = user.character_ownerships.all()
+        corporation_ids = {
+            character.character.corporation_id for character in ownerships
+        }
+        character_ownership_ids = {character.pk for character in ownerships}
+        request_query = self.select_related(
+            "blueprint__owner", "blueprint__owner__corporation"
+        ).filter(
+            (
+                Q(blueprint__owner__corporation__corporation_id__in=corporation_ids)
+                | Q(blueprint__owner__pk__in=character_ownership_ids)
+            )
+            & Q(closed_at=None)
+            & Q(status="IP")
+            & Q(fulfulling_user=user)
+        )
+        return request_query
+
+
 class RequestManager(models.Manager):
-    def requests_fulfillable_by_user(self, user: User) -> list:
-        corporation_ids = {
-            character.character.corporation_id
-            for character in user.character_ownerships.all()
-        }
+    def get_queryset(self) -> models.QuerySet:
+        return RequestQuerySet(self.model, using=self._db)
 
-        request_query = self.filter(
-            requestee_corporation__corporation_id__in=corporation_ids,
-            closed_at=None,
-            status="OP",
+    def select_related_default(self) -> models.QuerySet:
+        return self.select_related(
+            "blueprint",
+            "blueprint__owner",
+            "blueprint__eve_type",
+            "requesting_user__profile__main_character",
         )
-        return request_query
-
-    def requests_being_fulfilled_by_user(self, user: User) -> list:
-        corporation_ids = {
-            character.character.corporation_id
-            for character in user.character_ownerships.all()
-        }
-
-        request_query = self.filter(
-            requestee_corporation__corporation_id__in=corporation_ids,
-            closed_at=None,
-            status="IP",
-            fulfulling_user=user,
-        )
-        return request_query