+++ blueprints/tests/utils.py
@@ -41,6 +41,10 @@ def _generate_token(
             "esi-universe.read_structures.v1",
             "esi-corporations.read_blueprints.v1",
             "esi-assets.read_corporation_assets.v1",
+            "esi-characters.read_blueprints.v1",
+            "esi-assets.read_assets.v1",
+            "esi-industry.read_character_jobs.v1",
+            "esi-industry.read_corporation_jobs.v1",
         ]
     token = {
         "access_token": access_token,